+++ src/wago/module.go
@@ -56,8 +56,8 @@ type ImportSpec struct {
 	ValueType     ValueTypeDescriptor
 	HasValueType  bool
 	Mutable       bool
-	Min           int
-	Max           int
+	Min           uint64
+	Max           uint64
 	MemoryMin     uint64
 	MemoryMax     uint64
 	HasMax        bool
@@ -285,7 +285,7 @@ func (m *Module) Metadata() ModuleMetadata {
 		tables[i] = TableMetadata{Index: i, Type: c.tableElementType(i), ValueType: exact, HasValueType: exactErr == nil, Addr64: def.Addr64, Exports: tableExports[i]}
 		if imp, ok := c.tableImportAt(i); ok {
 			tables[i].ImportModule, tables[i].ImportName = splitImportKey(imp.Key)
-			tables[i].Min, tables[i].Max, tables[i].HasMax = uint64(imp.Min), uint64(imp.Max), imp.HasMax
+			tables[i].Min, tables[i].Max, tables[i].HasMax = imp.Min, imp.Max, imp.HasMax
 			continue
 		}
 		def = c.tableDef(i)
